yatube_api_project/requirements.txt
gunicorn==20.0.4
Django==3.2
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter
djoser==2.1.0
PyJWT==2.1.0
psycopg2-binary==2.9.3
requests==2.26.0
